Adds two keyboards shortcuts
Adds two keyboard shortcuts for use in the input text field. These are
primarily useful if the user is using the switcher to create a new file
(not open an existing file). The user can then start typing to view the
currently existing notes with similar names, the delete some of the search
terms or all of them to create a new note name, now that he has
investigated the current existing file names. The first keyboard shortcut
will delete all text in the input field. The second keyboard shortcut
deletes the word under the cursor.