Fix #171 (History corruption on close)
This PR drops all the history manipulation code and instead
does previews in-place in history (so the history of the
current tab never changes).  Basically, instead of relying
on the undocumented internals of Obsidian history, this
new version relies on the undocumented `popstate` field in
ViewState (that updates the current state without affecting
history).